+++ src/client/index.ts
@@ -4,4 +4,3 @@
 export * from "./core";
 export * from "./get";
 export * from "./post";
-export * from "./types";