+++ src/internal/account.ts
@@ -7,15 +7,11 @@
  * other namespaces and processes can access these methods without depending on the entire
  * account namespace and without having a dependency cycle error.
  */
-
 import { AptosConfig } from "../api/aptosConfig";
-import { AptosApiError, getAptosFullNode, paginateWithCursor } from "../client";
-import { AccountAddress, AccountAddressInput } from "../core/accountAddress";
-import { Account } from "../account";
-import { AnyPublicKey, Ed25519PublicKey, PrivateKey } from "../core/crypto";
-import { queryIndexer } from "./general";
+import { getAptosFullNode, paginateWithCursor } from "../client";
 import {
   AccountData,
+  AptosApiError,
   GetAccountCoinsDataResponse,
   GetAccountCollectionsWithOwnedTokenResponse,
   GetAccountOwnedTokensFromCollectionResponse,
@@ -31,6 +27,10 @@ import {
   TransactionResponse,
   WhereArg,
 } from "../types";
+import { AccountAddress, AccountAddressInput } from "../core/accountAddress";
+import { Account } from "../account";
+import { AnyPublicKey, Ed25519PublicKey, PrivateKey } from "../core/crypto";
+import { queryIndexer } from "./general";
 import {
   GetAccountCoinsCountQuery,
   GetAccountCoinsDataQuery,